BrieflyAI/requirements.txt
streamlit
fastapi
python-dotenv
beautifulsoup4
lxml
aiolimiter
langgraph
mcp
langchain-mcp-adapters
langchain-groq